+++ packages/agtree/src/converter/index.ts
@@ -2,6 +2,16 @@
  * @file Converter entry point.
  */
 export { FilterListConverter } from './filter-list';
-export { RawFilterListConverter } from './raw-filter-list';
+export { RawFilterListConverter, type FilterListConvertOptions } from './raw-filter-list';
 export { RuleConverter } from './rule';
 export { RawRuleConverter } from './raw-rule';
+export {
+    FilterListConversionResult,
+    type FilterListConversionError,
+} from './filter-list-conversion-result';
+export {
+    type ConversionSourceMap,
+    conversionSourceMapValidator,
+    createEmptyConversionSourceMap,
+} from './source-map';
+export { CONVERTER_PARSE_OPTIONS } from './parse-options';